+++ simulation/traffic_simulator/src/entity/pedestrian_entity.cpp
@@ -26,17 +26,22 @@ namespace entity
 {
 PedestrianEntity::PedestrianEntity(
   const std::string & name, const openscenario_msgs::msg::PedestrianParameters & params)
-: EntityBase(params.pedestrian_category, name), parameters(params)
+: EntityBase(params.pedestrian_category, name),
+  parameters(params),
+  plugin_name("behavior_tree_plugin/PedestrianBehaviorTree"),
+  loader_(pluginlib::ClassLoader<entity_behavior::BehaviorPluginBase>(
+    "traffic_simulator", "entity_behavior::BehaviorPluginBase")),
+  behavior_plugin_ptr_(loader_.createSharedInstance(plugin_name))
 {
   entity_type_.type = openscenario_msgs::msg::EntityType::PEDESTRIAN;
-  tree_ptr_ = std::make_shared<entity_behavior::pedestrian::BehaviorTree>();
-  tree_ptr_->setValueToBlackBoard("pedestrian_parameters", parameters);
+  behavior_plugin_ptr_->configure();
+  behavior_plugin_ptr_->setPedestrianParameters(parameters);
 }
 
 void PedestrianEntity::requestAssignRoute(
   const std::vector<openscenario_msgs::msg::LaneletPose> & waypoints)
 {
-  tree_ptr_->setRequest("follow_lane");
+  behavior_plugin_ptr_->setRequest("follow_lane");
   if (!status_) {
     return;
   }
@@ -60,12 +65,12 @@ void PedestrianEntity::requestAssignRoute(const std::vector<geometry_msgs::msg::
   requestAssignRoute(route);
 }
 
-void PedestrianEntity::requestWalkStraight() { tree_ptr_->setRequest("walk_straight"); }
+void PedestrianEntity::requestWalkStraight() { behavior_plugin_ptr_->setRequest("walk_straight"); }
 
 void PedestrianEntity::requestAcquirePosition(
   const openscenario_msgs::msg::LaneletPose & lanelet_pose)
 {
-  tree_ptr_->setRequest("follow_lane");
+  behavior_plugin_ptr_->setRequest("follow_lane");
   if (!status_) {
     return;
   }
@@ -87,7 +92,7 @@ void PedestrianEntity::requestAcquirePosition(const geometry_msgs::msg::Pose & m
 
 void PedestrianEntity::cancelRequest()
 {
-  tree_ptr_->setRequest("none");
+  behavior_plugin_ptr_->setRequest("none");
   route_planner_ptr_->cancelGoal();
 }
 
@@ -105,20 +110,20 @@ void PedestrianEntity::onUpdate(double current_time, double step_time)
   if (current_time < 0) {
     updateEntityStatusTimestamp(current_time);
   } else {
-    tree_ptr_->setValueToBlackBoard("other_entity_status", other_status_);
-    tree_ptr_->setValueToBlackBoard("entity_type_list", entity_type_list_);
-    tree_ptr_->setValueToBlackBoard("entity_status", status_.get());
+    behavior_plugin_ptr_->setOtherEntityStatus(other_status_);
+    behavior_plugin_ptr_->setEntityTypeList(entity_type_list_);
+    behavior_plugin_ptr_->setEntityStatus(status_.get());
     target_speed_planner_.update(status_->action_status.twist.linear.x);
-    tree_ptr_->setValueToBlackBoard("target_speed", target_speed_planner_.getTargetSpeed());
+    behavior_plugin_ptr_->setTargetSpeed(target_speed_planner_.getTargetSpeed());
     if (status_->lanelet_pose_valid) {
       auto route = route_planner_ptr_->getRouteLanelets(status_->lanelet_pose);
-      tree_ptr_->setValueToBlackBoard("route_lanelets", route);
+      behavior_plugin_ptr_->setRouteLanelets(route);
     } else {
       std::vector<std::int64_t> empty = {};
-      tree_ptr_->setValueToBlackBoard("route_lanelets", empty);
+      behavior_plugin_ptr_->setRouteLanelets(empty);
     }
-    tree_ptr_->tick(current_time, step_time);
-    auto status_updated = tree_ptr_->getUpdatedStatus();
+    behavior_plugin_ptr_->update(current_time, step_time);
+    auto status_updated = behavior_plugin_ptr_->getUpdatedStatus();
     if (status_updated.lanelet_pose_valid) {
       auto following_lanelets =
         hdmap_utils_ptr_->getFollowingLanelets(status_updated.lanelet_pose.lanelet_id);