+++ simulation/traffic_simulator/src/entity/vehicle_entity.cpp
@@ -29,10 +29,14 @@ VehicleEntity::VehicleEntity(
   const std::string & name, const openscenario_msgs::msg::VehicleParameters & params)
 : EntityBase(params.vehicle_category, name),
   parameters(params),
-  tree_ptr_(std::make_shared<entity_behavior::vehicle::BehaviorTree>())
+  plugin_name("behavior_tree_plugin/VehicleBehaviorTree"),
+  loader_(pluginlib::ClassLoader<entity_behavior::BehaviorPluginBase>(
+    "traffic_simulator", "entity_behavior::BehaviorPluginBase")),
+  behavior_plugin_ptr_(loader_.createSharedInstance(plugin_name))
 {
   entity_type_.type = openscenario_msgs::msg::EntityType::VEHICLE;
-  tree_ptr_->setValueToBlackBoard("vehicle_parameters", parameters);
+  behavior_plugin_ptr_->configure();
+  behavior_plugin_ptr_->setVehicleParameters(parameters);
 }
 
 void VehicleEntity::requestAssignRoute(
@@ -76,13 +80,13 @@ void VehicleEntity::requestAcquirePosition(const geometry_msgs::msg::Pose & map_
 
 void VehicleEntity::requestLaneChange(const std::int64_t to_lanelet_id)
 {
-  tree_ptr_->setRequest("lane_change");
-  tree_ptr_->setValueToBlackBoard("to_lanelet_id", to_lanelet_id);
+  behavior_plugin_ptr_->setRequest("lane_change");
+  behavior_plugin_ptr_->setToLaneletId(to_lanelet_id);
 }
 
 void VehicleEntity::cancelRequest()
 {
-  tree_ptr_->setRequest("none");
+  behavior_plugin_ptr_->setRequest("none");
   route_planner_ptr_->cancelGoal();
 }
 
@@ -100,20 +104,20 @@ void VehicleEntity::onUpdate(double current_time, double step_time)
   if (current_time < 0) {
     updateEntityStatusTimestamp(current_time);
   } else {
-    tree_ptr_->setValueToBlackBoard("other_entity_status", other_status_);
-    tree_ptr_->setValueToBlackBoard("entity_type_list", entity_type_list_);
-    tree_ptr_->setValueToBlackBoard("entity_status", status_.get());
+    behavior_plugin_ptr_->setOtherEntityStatus(other_status_);
+    behavior_plugin_ptr_->setEntityTypeList(entity_type_list_);
+    behavior_plugin_ptr_->setEntityStatus(status_.get());
     target_speed_planner_.update(status_->action_status.twist.linear.x);
-    tree_ptr_->setValueToBlackBoard("target_speed", target_speed_planner_.getTargetSpeed());
+    behavior_plugin_ptr_->setTargetSpeed(target_speed_planner_.getTargetSpeed());
     if (status_->lanelet_pose_valid) {
-      tree_ptr_->setValueToBlackBoard(
-        "route_lanelets", route_planner_ptr_->getRouteLanelets(status_->lanelet_pose));
+      behavior_plugin_ptr_->setRouteLanelets(
+        route_planner_ptr_->getRouteLanelets(status_->lanelet_pose));
     } else {
       std::vector<std::int64_t> empty = {};
-      tree_ptr_->setValueToBlackBoard("route_lanelets", empty);
+      behavior_plugin_ptr_->setRouteLanelets(empty);
     }
-    tree_ptr_->tick(current_time, step_time);
-    auto status_updated = tree_ptr_->getUpdatedStatus();
+    behavior_plugin_ptr_->update(current_time, step_time);
+    auto status_updated = behavior_plugin_ptr_->getUpdatedStatus();
     if (status_updated.lanelet_pose_valid) {
       auto following_lanelets =
         hdmap_utils_ptr_->getFollowingLanelets(status_updated.lanelet_pose.lanelet_id);